Upload Pb2Chts Debugreport
Lots of people complaining about the script not working on their device, and i cant find the specific root problem, and there is a lot of devices/platforms that may not work.

This script generates a report, so the user can just send that with the issue report.

Its still pretty basic and unstable though, so crash log for this script also appreciated :+1: